move {StoryView,CommentView}::construct_help_view into help_view module
